Fix extraction of multiple nullbytes in a row
The regex used for matching null bytes without a leading slash consumed
a character that was no longer available for further matching. This is
a problem as soon as null bytes occur next to each other.

In Python, we can use a negative lookbehind assertion to match all
cases. For sed, a second call with the same pattern does the trick.

Fixes GH-1