Add option for recursive unlink
This changeset adds an option (--recursive-unlink) that can be used when
extracting a ttar archive. If the option is set and the archive contains
a file where a directory already exists, that directory will be removed
recursively, even if it is not empty).

This option is modeled after an option of the same name in GNU tar.